requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
django-model-utils==4.5.1
djangorestframework==3.15.2
Markdown==3.7
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1